Fixed time-based vars that were still float instead of double.

+++ Source/Client/Game/Private/Game.cpp
@@ -187,7 +187,7 @@ World& Game::getWorld()
     return world;
 }
 
-float Game::getAccumulatedTime()
+double Game::getAccumulatedTime()
 {
     return accumulatedTime;
 }
