+++ Source/Client/Utility/Public/Network.h
@@ -94,7 +94,7 @@ private:
     static constexpr unsigned int SERVER_PORT = 41499;
 
     /** How long we should wait before considering the server to be timed out. */
-    static constexpr float TIMEOUT_S = NETWORK_TICK_INTERVAL_S * 2;
+    static constexpr double TIMEOUT_S = NETWORK_TICK_INTERVAL_S * 2;
 
     std::shared_ptr<Peer> server;
